// src/app/auth/interfaces/IigTbFactNomina.ts
export interface IigTbFactNomina {
    acumulaDecCuarto: string;
    acumulaDecTercero: string;
    apellidosNombres: string;
    aporteIndividual: number;
    aporteIndividualAdi: number;
    aportePatronal: number;
    bonoGeografico: string;
    cesantiaFreserva: number;
    codigoActividadLaboral: string;
    codigoDenominacionPuesto: string;
    codigoEscalaOcupacional: string;
    codigoEstadoPuesto: string;
    codigoEstadoServidor: string;
    codigoInstitucion: string;
    codigoModalidadLaboral: string;
    codigoNivelOcupacional: string;
    codigoRegimenLaboral: string;
    codigoSucursal: string;
    codigoUnidadOrganizacional: string;
    comisionServicio: string;
    comisionServicioTipo: string;
    ejercicioFiscal: number;
    enlaceConcepto: string;
    enlaceEstructuraProgramatica: string;
    enlaceGrupo: string;
    estructuraOrganica: string;
    fechaFin: Date;
    fechaInicio: Date;
    fechaNacimiento: Date;
    fechaReporte: Date;
    fondoReserva: string;
    fuente: string;
    genero: string;
    gradoEscalaOcupacional: number;
    identificadorDistributivo: number;
    nombreActividadLaboral: string;
    nombreCanton: string;
    nombreDenominacionPuesto: string;
    nombreEscalaOcupacional: string;
    nombreEstadoPuesto: string;
    nombreEstadoServidor: string;
    nombreInstitucion: string;
    nombreModalidadLaboral: string;
    nombreNivelOcupacional: string;
    nombreProvincia: string;
    nombreRegimenLaboral: string;
    nombreUnidadOrganizacional: string;
    numeroDocumento: string;
    partidaIndividual: number;
    porcentajeAportePatronalAdi: number;
    porcentajeIece: number;
    porcentajeSecap: number;
    rmuEscalaOcupacional: number;
    rmuEscalaOcupacionalMaximo: number;
    rmuPuesto: number;
    rmuSobrevalorado: string;
    rucPatronal: string;
    tipoDocumento: string;
}